requirements.txt
paho-mqtt==1.5.0
rhasspy-hermes~=0.6.0
rhasspy-silence~=0.4.0
openai==0.27.8
rhasspy-asr
